+++ src/main/java/org/kamiblue/client/mixin/client/MixinStateImplementation.java
@@ -3,16 +3,20 @@ package org.kamiblue.client.mixin.client;
 import net.minecraft.block.Block;
 import net.minecraft.block.state.BlockStateContainer;
 import net.minecraft.entity.Entity;
+import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import org.kamiblue.client.module.modules.movement.Jesus;
+import org.kamiblue.client.module.modules.render.Xray;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import java.util.List;
 
@@ -23,5 +27,12 @@ public class MixinStateImplementation {
     @Inject(method = "addCollisionBoxToList", at = @At("HEAD"), cancellable = true)
     public void addCollisionBoxToList(World worldIn, BlockPos pos, AxisAlignedBB entityBox, List<AxisAlignedBB> collidingBoxes, Entity entityIn, boolean isActualState, CallbackInfo ci) {
         Jesus.handleAddCollisionBoxToList(pos, block, entityIn, collidingBoxes);
+    }
+
+    @Inject(method = "shouldSideBeRendered", at = @At("HEAD"), cancellable = true)
+    public void shouldSideBeRendered(IBlockAccess blockAccess, BlockPos pos, EnumFacing facing, CallbackInfoReturnable<Boolean> ci) {
+        if (Xray.shouldReplace(blockAccess.getBlockState(pos.offset(facing)))) {
+            ci.setReturnValue(true);
+        }
     }
 }
