--- src/main/java/org/kamiblue/client/mixin/client/render/MixinStateImplementation.java
@@ -1,23 +1,0 @@
-package org.kamiblue.client.mixin.client.render;
-
-import net.minecraft.block.state.BlockStateContainer;
-import net.minecraft.util.EnumFacing;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.IBlockAccess;
-import org.kamiblue.client.module.modules.render.Xray;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-
-@Mixin(BlockStateContainer.StateImplementation.class)
-public class MixinStateImplementation {
-
-    @Inject(method = "shouldSideBeRendered", at = @At("HEAD"), cancellable = true)
-    public void shouldSideBeRendered(IBlockAccess blockAccess, BlockPos pos, EnumFacing facing, CallbackInfoReturnable<Boolean> ci) {
-        if (Xray.shouldReplace(blockAccess.getBlockState(pos.offset(facing)))) {
-            ci.setReturnValue(true);
-        }
-    }
-
-}
