+++ src/main/java/org/kamiblue/client/mixin/client/world/MixinBlockModelRenderer.java
@@ -1,4 +1,4 @@
-package org.kamiblue.client.mixin.client.render;
+package org.kamiblue.client.mixin.client.world;
 
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.renderer.BlockModelRenderer;
@@ -14,12 +14,10 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(BlockModelRenderer.class)
 public class MixinBlockModelRenderer {
-
     @Inject(method = "renderModel(Lnet/minecraft/world/IBlockAccess;Lnet/minecraft/client/renderer/block/model/IBakedModel;Lnet/minecraft/block/state/IBlockState;Lnet/minecraft/util/math/BlockPos;Lnet/minecraft/client/renderer/BufferBuilder;ZJ)Z", at = @At("HEAD"), cancellable = true)
     public void renderModel(IBlockAccess worldIn, IBakedModel modelIn, IBlockState stateIn, BlockPos posIn, BufferBuilder buffer, boolean checkSides, long rand, CallbackInfoReturnable<Boolean> ci) {
         if (Xray.shouldReplace(stateIn)) {
             ci.cancel();
         }
     }
-
 }
