+++ src/main/resources/mixins.kami.json
@@ -49,11 +49,7 @@
     "player.MixinEntityPlayer",
     "player.MixinEntityPlayerSP",
     "player.MixinPlayerControllerMP",
-    "render.MixinBlockFluidRenderer",
-    "render.MixinBlockModelRenderer",
-    "render.MixinBlock",
     "render.MixinDebugRendererChunkBorder",
-    "render.MixinEntity",
     "render.MixinEntityRenderer",
     "render.MixinFontRenderer",
     "render.MixinItemRenderer",
@@ -69,13 +65,15 @@
     "render.MixinRenderLivingBase",
     "render.MixinRenderManager",
     "render.MixinRenderPlayer",
-    "render.MixinStateImplementation",
     "render.MixinTileEntityRendererDispatcher",
     "render.MixinTileEntitySignRenderer",
     "render.MixinTileRendererDispatcher",
     "render.MixinViewFrustum",
     "render.MixinVisGraph",
+    "world.MixinBlock",
+    "world.MixinBlockFluidRenderer",
     "world.MixinBlockLiquid",
+    "world.MixinBlockModelRenderer",
     "world.MixinBlockSoulSand",
     "world.MixinBlockWeb",
     "world.MixinWorld"
